+++ tests/CommandTest.php
@@ -8,6 +8,7 @@ use Throwable;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
+use Yiisoft\Db\Expression\JsonExpression;
 use Yiisoft\Db\Schema\SchemaInterface;
 use Yiisoft\Db\Sqlite\Tests\Support\TestTrait;
 use Yiisoft\Db\Tests\Common\CommonCommandTest;
@@ -494,5 +495,39 @@ final class CommandTest extends CommonCommandTest
 
         $this->assertSame('sqlite::memory:', $db->getDriver()->getDsn());
         $this->assertSame(['main'], $command->showDatabases());
+    }
+
+    public function testJsonTable(): void
+    {
+        $db = $this->getConnection();
+        $command = $db->createCommand();
+
+        if ($db->getTableSchema('json_table', true) !== null) {
+            $command->dropTable('json_table')->execute();
+        }
+
+        $command->createTable('json_table', [
+            'id' => SchemaInterface::TYPE_PK,
+            'json_col' => SchemaInterface::TYPE_JSON,
+        ])->execute();
+
+        $command->insert('json_table', ['id' => 1, 'json_col' => ['a' => 1, 'b' => 2]])->execute();
+        $command->insert('json_table', ['id' => 2, 'json_col' => new JsonExpression(['c' => 3, 'd' => 4])])->execute();
+
+        $tableSchema = $db->getTableSchema('json_table', true);
+        $this->assertNotNull($tableSchema);
+        $this->assertSame('json_col', $tableSchema->getColumn('json_col')->getName());
+        $this->assertSame('json', $tableSchema->getColumn('json_col')->getType());
+        $this->assertSame('json', $tableSchema->getColumn('json_col')->getDbType());
+
+        $this->assertSame(
+            '{"a":1,"b":2}',
+            $command->setSql('SELECT `json_col` FROM `json_table` WHERE `id`=1')->queryScalar(),
+        );
+
+        $this->assertSame(
+            '{"c":3,"d":4}',
+            $command->setSql('SELECT `json_col` FROM `json_table` WHERE `id`=2')->queryScalar(),
+        );
     }
 }