+++ tests/Provider/SchemaProvider.php
@@ -222,6 +222,32 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                         'scale' => null,
                         'defaultValue' => 0b1000_0010, // 130
                     ],
+                    'json_col' => [
+                        'type' => 'json',
+                        'dbType' => 'json',
+                        'phpType' => 'array',
+                        'primaryKey' => false,
+                        'allowNull' => false,
+                        'autoIncrement' => false,
+                        'enumValues' => null,
+                        'size' => null,
+                        'precision' => null,
+                        'scale' => null,
+                        'defaultValue' => ['number' => 10],
+                    ],
+                    'json_text_col' => [
+                        'type' => 'json',
+                        'dbType' => 'json',
+                        'phpType' => 'array',
+                        'primaryKey' => false,
+                        'allowNull' => true,
+                        'autoIncrement' => false,
+                        'enumValues' => null,
+                        'size' => null,
+                        'precision' => null,
+                        'scale' => null,
+                        'defaultValue' => null,
+                    ],
                 ],
                 'tableName' => 'type',
             ],
@@ -405,7 +431,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
         $constraints['2: unique'][2][0]->name(AnyValue::getInstance());
         $constraints['2: index'][2][2]->name(AnyValue::getInstance());
 
-        $constraints['3: foreign key'][2][0]->name(null);
+        $constraints['3: foreign key'][2][0]->name('0');
         $constraints['3: index'][2] = [];
 
         $constraints['4: primary key'][2]->name(null);